# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Topos OSS plane",
    "description": "The self-hostable Topos plane — workspace-credential writes AND reads (one Bearer credential per enrolled device; membership is the authorization), plus the enrollment + governance surface. Every returned protocol outcome of an op_id-carrying write rides in a 200 body (the canonical JsonEnvelope + receipt); non-2xx is reserved for transport/auth/integrity faults.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/v1/admin-claim": {
      "post": {
        "tags": [
          "enrollment"
        ],
        "operationId": "admin_claim",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The admin-claim receipt — OK stands up the workspace + seats the first owner; DENIED the flat error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body or device public key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/authorize": {
      "post": {
        "tags": [
          "enrollment"
        ],
        "operationId": "start_device_auth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceAuthorizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The device-authorization grant (device_code / user_code / verification_uri(_complete)) plus the plane block (API base + posture + method, no trust root) every start now carries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceAuthorizeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body, device public key, workspace name, or intent/workspace combination.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "A standup start on a plane that does not offer it (self-host).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/device/token": {
      "post": {
        "tags": [
          "enrollment"
        ],
        "operationId": "poll_device_auth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The poll status (pending/slow_down/denied/expired/granted + the grant on granted).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such device code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/enroll/passcode/confirm": {
      "post": {
        "tags": [
          "enrollment"
        ],
        "operationId": "complete_passcode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasscodeConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The confirmation status (confirmed/wrong_code/expired/too_many_attempts).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasscodeConfirmResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No live session for that user code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/enroll/verify/{user_code}": {
      "get": {
        "tags": [
          "enrollment"
        ],
        "operationId": "read_verification_context",
        "parameters": [
          {
            "name": "user_code",
            "in": "path",
            "description": "The user code shown by `device/authorize`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The verification-page disclosure (machine name, fingerprint, workspace, offered skills).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationContextResponse"
                }
              }
            }
          },
          "404": {
            "description": "No live session for that user code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/login": {
      "post": {
        "tags": [
          "enrollment"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRedeemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The login receipt — OK carries the LoginData (the proven identity + one re-minted credential, or a blocked marker, per confirmed seat); DENIED the flat error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body or device public key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{ws}/devices": {
      "post": {
        "tags": [
          "enrollment"
        ],
        "operationId": "redeem",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace id (the grant is authoritative for the workspace).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The redeem receipt — OK carries the registered device + minted read creds; DENIED the flat error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body or key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "governance"
        ],
        "operationId": "revoke_device",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace id (REST sugar; the authoritative workspace_id rides the body).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "`Bearer <workspace credential>` — the acting device's credential (owner, or the target's own principal).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The revoke receipt — OK on success (instant revoke), DENIED (a 200) on a role denial.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body or identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal store fault.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminClaimRequest": {
        "type": "object",
        "description": "`POST /v1/admin-claim` body — consume a one-time self-host claim token to stand up a workspace + seat its\nfirst owner. The server re-derives the device key id from `device_public_key`.",
        "required": [
          "claim_token",
          "device_public_key",
          "display_name"
        ],
        "properties": {
          "claim_token": {
            "type": "string",
            "description": "The one-time admin-claim token."
          },
          "device_public_key": {
            "type": "string",
            "description": "The claiming device's raw 32-byte public key, base64url-unpadded."
          },
          "display_name": {
            "type": "string",
            "description": "The display name for the standing-up workspace."
          }
        }
      },
...
      "DeviceAuthorizeRequest": {
        "type": "object",
        "description": "`POST /v1/device/authorize` body — begin an RFC-8628 device-authorization flow. With a `workspace`\n(the address name; the default `enroll` intent) the device enrolls toward that workspace; with\n`intent = \"standup\"` — or neither field — it starts a STANDUP session (no workspace yet; hosted\nplanes only); with `intent = \"login\"` it starts a workspace-less LOGIN session whose grant redeems\nat `POST /v1/login`. The server SERVER-derives the device key id from `device_public_key` (a\nclient-asserted id is never trusted).",
        "required": [
          "device_public_key",
          "machine_name"
        ],
        "properties": {
          "device_public_key": {
            "type": "string",
            "description": "The device's raw 32-byte public key, base64url-unpadded. The server derives the device key id from\nit (never a client-asserted id) and binds it to the enrollment session."
          },
          "intent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SessionIntent",
                "description": "The session intent. Absent defaults to `enroll` when a workspace is named, `standup` otherwise."
              }
            ]
          },
          "machine_name": {
            "type": "string",
            "description": "A human-readable machine name shown on the verification page (a confused-deputy guard, not authority)."
          },
          "workspace": {
            "type": [
              "string",
              "null"
            ],
            "description": "The workspace ADDRESS name an enroll session targets (`topos.sh/<name>` minus the origin).\nWhether the name exists is never disclosed on this route: an unknown name runs the same flow to\nthe same uniform denial a not-yours workspace gets."
          }
        }
      },
      "DeviceAuthorizeResponse": {
        "type": "object",
        "description": "`POST /v1/device/authorize` response — the RFC-8628 device-authorization grant (the names mirror the RFC).",
        "required": [
          "device_code",
          "user_code",
          "verification_uri",
          "expires_in",
          "interval"
        ],
        "properties": {
          "device_code": {
            "type": "string",
            "description": "The SECRET device code the client polls `device/token` with."
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "description": "The session lifetime, in seconds (RFC-8628 `expires_in`).",
            "minimum": 0
          },
          "interval": {
            "type": "integer",
            "format": "int64",
            "description": "The minimum poll interval, in seconds (RFC-8628 `interval`).",
            "minimum": 0
          },
          "plane": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BootstrapPlane",
                "description": "The plane block a STANDUP start carries (base URL, deployment posture, enrollment method) — a\nstandup device has no `/i/` bootstrap to learn these from. Absent on an enroll start."
              }
            ]
          },
          "user_code": {
            "type": "string",
            "description": "The opaque code identifying this device-auth session; it rides inside `verification_uri_complete`\n(the human clicks the URL — it is not typed)."
          },
          "verification_uri": {
            "type": "string",
            "description": "The verification URL a human visits to approve the session."
          },
          "verification_uri_complete": {
            "type": [
              "string",
              "null"
            ],
            "description": "The verification URL with the user code already embedded (RFC-8628 `verification_uri_complete`) —\nthe one link to open; a client uses it VERBATIM when present."
          }
        }
      },
...
      "DeviceTokenRequest": {
        "type": "object",
        "description": "`POST /v1/device/token` body — poll a device-authorization session for its grant.",
        "required": [
          "device_code"
        ],
        "properties": {
          "device_code": {
            "type": "string",
            "description": "The SECRET device code from `device/authorize`."
          }
        }
      },
      "DeviceTokenResponse": {
        "type": "object",
        "description": "`POST /v1/device/token` response — the poll `status`, plus the opaque single-use enrollment `grant` ONLY\nwhen `status` is `granted`. A re-poll of a confirmed session re-derives the SAME grant (idempotent issue).",
        "required": [
          "status"
        ],
        "properties": {
          "grant": {
            "type": [
              "string",
              "null"
            ],
            "description": "The opaque single-use enrollment grant — present ONLY when `status` is `granted` (the redeem credential)."
          },
          "status": {
            "$ref": "#/components/schemas/DeviceTokenStatus",
            "description": "The poll status."
          },
          "workspace": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeviceTokenWorkspace",
                "description": "The granted session's workspace context — present when `status` is `granted`."
              }
            ]
          }
        }
      },
...
      "JsonEnvelope": {
        "type": "object",
        "description": "The one common envelope every verb emits. Never prompts; prose (TTY) is rendered from the\nSAME typed value where practical.",
        "required": [
          "schema_version",
          "command",
          "ok"
        ],
        "properties": {
          "command": {
            "type": "string",
            "description": "The verb that produced this (`add`, `follow`, `pull`, `list`, `publish`, …)."
          },
          "data": {
            "description": "Command-specific payload (`{}` when empty)."
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WireError",
                "description": "The actionable failure detail (when `ok == false`)."
              }
            ]
          },
          "next_actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NextAction"
            },
            "description": "Stable, machine-actionable next steps — each carries a complete argv."
          },
          "ok": {
            "type": "boolean"
          },
          "receipt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Receipt",
                "description": "The durable idempotency record — present on **every** terminal op, including failures."
              }
            ]
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "description": "Always `1` for this contract version (the schema pins it `const`).",
            "minimum": 0
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}